Refactor client command handling
